requirements.txt
fastapi[all]
pydantic
python-dotenv
mistralai  # Replace with the actual version if different
uvicorn